+++ src/core/tabs/tabs-demo.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-export default [...Array(6)].map((_, idx) => ({
+export default [...Array(4)].map((_, idx) => ({
   label: `切换到内容${idx + 1}`,
   children: (
     <div className="p10" key={idx}>内容 {idx + 1}</div>